requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argilla==1.4.0
asgiref==3.6.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
backoff==2.2.1
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==15.0.1
commonmark==0.9.1
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
Cython==0.29.33
dataclasses-json==0.5.7
Deprecated==1.2.13
dnspython==2.3.0
effdet==0.3.0
et-xmlfile==1.1.0
fastapi==0.93.0
filelock==3.9.0
firebase-admin==6.1.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Login==0.6.2
flatbuffers==23.3.3
fonttools==4.39.0
frozenlist==1.3.3
gcloud==0.18.3
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-cloud-vision==3.4.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
greenlet==2.0.2
grpcio==1.51.3
grpcio-status==1.51.3
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.21.0
httpx==0.23.3
huggingface-hub==0.13.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
install==1.3.5
iopath==0.1.10
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsons==1.6.3
jws==0.1.3
jwt==1.3.1
kiwisolver==1.4.4
langchain==0.0.104
layoutparser==0.3.4
loguru==0.6.0
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.23.5
oauth2client==4.1.3
omegaconf==2.3.0
onnxruntime==1.14.1
openai==0.27.1
opencv-python==4.6.0.66
openpyxl==3.1.1
packaging==23.0
pandas==1.5.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.8.0
Pillow==9.4.0
pinecone-client==2.2.1
poppler-utils==0.1.0
portalocker==2.7.0
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.5
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==3.0.9
pypdf==3.5.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-jwt==2.0.1
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==1.5.0
rich==13.0.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.25.0
sympy==1.11.1
tenacity==8.2.2
timm==0.6.12
tokenizers==0.13.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
transformers==4.26.1
typing-inspect==0.8.0
typing_extensions==4.5.0
typish==1.9.3
unstructured==0.5.3
unstructured-inference==0.2.7
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.21.0
Wand==0.6.11
Werkzeug==2.2.3
wrapt==1.14.1
XlsxWriter==3.0.8
yarl==1.8.2
zipp==3.15.0
